ENH: Reorganized Code
