+++ Supervised/image_info.csv
