+++ Unsupervised/Encoder.py
